# Samples/Asp.net Core - Support Microsoft/Asp.net Core - Support Microsoft/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Microsoft Support Plugin",
    "description": "Returns the first response from Microsoft Support to the user when they have a question regarding one of Microsoft products",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://ab431b30cf6b.ngrok.app"
    }
  ],
  "paths": {
    "/MicrosoftSupport": {
      "get": {
        "operationId": "GetMicrosoftSupportResponse",
        "summary": "Get the Microsoft Support response for a question give in query parameters",
        "parameters": [
          {
            "name": "question",
            "in": "query",
            "description": "Represents the question to ask Microsoft support in natural language as asked by the user",
            "required": true,
            "schema": {
              "type": "string",
              "example": "how can i use Microsoft Teams"
            }
          },
          {
            "name": "originalLanguageFromIsoCode",
            "in": "query",
            "description": "Represents the language used by the user for the question, the response must be in same language",
            "required": true,
            "schema": {
              "type": "string",
              "example": "en-US"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/microsoftSupport"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "microsoftSupport": {
        "type": "object",
        "description": "The first response from Microsoft Support",
        "properties": {
          "title": {
            "type": "string",
            "description": "Support response title"
          },
          "linkEncoded": {
            "type": "string",
            "description": "Encode link placed on the title of the response from Microsoft Support"
          },
          "link": {
            "type": "string",
            "description": "Decoded link placed on Microsoft Support response title"
          },
          "description": {
            "type": "string",
            "description": "Description sent by Microsoft Support"
          },
          "appliesTo": {
            "type": "string",
            "description": "Microsoft application related to the response (may be null)"
          },
          "isoCodeLanguage": {
            "type": "string",
            "description": "Represente the language must be use for GPT response"
          }
        }
      }
    }
  }
}